completed refactoring of syscalls

+++ kernel/syscalls/syscalls.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 #include <stddef.h>
 
-void create_task(void (*code)(void *), void* args, uint32_t priority);
-void kcreate_task(void (*code)(void *), void *args, size_t priority);
+void create_task(void (*code)(void *), void* args, uint8_t priority);
+void kcreate_task(void (*code)(void *), void *args, uint8_t priority);
 void task_switch();
 
 #endif
